Switch to using references instead of copies
Rust is pass by value by default, so that means the Config param to push_msg was being copied.
this commit changes it so that the Config is passed by reference.

Additionally, the vector of parameters that are sent to Pushover don't need to be re-allocated, so
this commit makes those &strs instead of Strings.